WebApp/requirements.txt
cachetools==4.1.1
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
Flask==1.1.2
google-api-core==1.22.4
google-auth==1.22.1
google-cloud-core==1.4.3
google-cloud-storage==1.31.1
google-cloud-texttospeech==2.2.0
google-cloud-translate==3.0.1
google-cloud-vision==1.0.0
google-crc32c==1.0.0
google-resumable-media==1.1.0
googleapis-common-protos==1.52.0
grpcio==1.32.0
gunicorn==20.0.4
idna==2.10
itsdangerous==1.1.0
Jinja2==2.11.2
libcst==0.3.12
MarkupSafe==1.1.1
mypy-extensions==0.4.3
proto-plus==1.10.1
protobuf==3.13.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
rsa==4.6
six==1.15.0
typing-extensions==3.7.4.3
typing-inspect==0.6.0
urllib3==1.25.10
Werkzeug==1.0.1
numpy==1.19.3